+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Introduce `elasticstack_kibana_import_saved_objects` resource as an additive only way to manage Kibana saved objects ([#343](https://github.com/elastic/terraform-provider-elasticstack/pull/343)).
 - Add support for Terraform Plugin Framework ([#343](https://github.com/elastic/terraform-provider-elasticstack/pull/343)).
+- Fix Fleet Agent Policy monitoring settings ([#1](https://github.com/elastic/terraform-provider-elasticstack/pull/1))
 
 ## [0.9.0] - 2023-10-09